naja par chlini sache
git-svn-id: http://sbb.frickler.ch/svn/bfh.swed.monitor@7 3ec84a3b-7d30-49ee-bd3f-accf520b4063

+++ PatientMonitor/src/patientmonitor/definition/Device.java
@@ -36,6 +36,6 @@ public interface Device {
      * 
      * @return Returns the temperature
      */
-    public Integer performMeasure() throws NoPatientAssignedException;
+    public float performMeasure() throws NoPatientAssignedException;
 
 }
